Add files via upload
#Stopwatch Web Application :

This repository contains the code for a fully functional stopwatch web application built using HTML, CSS, and JavaScript. The application allows users to start, stop, and reset the timer with a simple and intuitive interface. Ideal for anyone looking to learn or improve their front-end development skills.

# Features
- Start, stop, and reset functionality
- Responsive and user-friendly design
- Accurate time tracking
- Clear and easy-to-read display

#Technologies Used
- HTML
- CSS
- JavaScript

Feel free to clone, fork, or contribute to this project!